code/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
apparmor==4.1.0
asgiref==3.8.1
attrs==24.1.0
autocommand==2.2.2
blinker==1.9.0
Brlapi==0.8.6
Brotli==1.1.0
btrfsutil==6.14
build==1.2.2
cachetools==5.5.0
cairocffi==1.6.1
capstone==5.0.6
certifi==2025.4.26
cffi==1.17.1
cfractions==2.4.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
ConfigArgParse==1.7.1
coverage==7.8.2
cramjam==2.10.0
cryptography==44.0.3
Cython==3.1.1
dbus-python==1.4.0
defusedxml==0.7.1
dirty-equals==0.9.0
distlib==0.3.9
distro-info==1.11
Django==5.1.6
django-ninja==1.3.0
django-ninja-simple-jwt==0.6.1
djangorestframework==3.16.0
docopt==0.6.2
docutils==0.21.2
editables==0.5
fastanime==2.8.7
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.3
ground==9.0.0
hatchling==1.27.0
httpie==3.2.4
hypothesis==6.131.23
idna==3.10
iniconfig==2.1.0
InquirerPy==0.3.4
installer==0.7.0
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyutils==0.6
lark==1.2.2
leechcorepyc==2.22.2
legacy-cgi==2.6.3
LibAppArmor==4.1.0
libfdt==1.7.2
lit==19.1.7.dev0
locus==10.0.0
locust==2.32.10
louis==3.33.0
Mako==1.3.10.dev0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
meson==1.8.1
more-itertools==10.7.0
msgpack==1.1.1
multidict==6.4.4
mysqlclient==2.2.7
netaddr==1.3.0
netsnmp-python==1.0a1
nose==1.3.7
numpy==2.2.6
ordered-set==4.1.0
packaging==25.0
pathspec==0.12.1
pefile==2024.8.26
pfzy==0.3.4
pillow==10.4.0
pipenv==2025.0.3
platformdirs==4.3.6
pluggy==1.6.0
plyer==2.1.0
poetry-core==2.1.3
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
pyaml==24.12.0
pycairo==1.28.0
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyGObject==3.52.3
PyJWT==2.10.1
pypresence==4.3.0
pyproject_hooks==1.2.0
PySide6==6.9.0
PySocks==1.7.1
pytest==8.3.5
python-dotenv==1.1.0
python-snappy==0.7.2
pytz==2025.2
pyxdg==0.28
PyYAML==6.0.2
pyzmq==27.0.0
qtile==0.31.1.dev0+g8666bfc8.d20250312
RapidFuzz==3.12.2
referencing==0.35.1
Reflector==2023.6.28.0.36.1
reprit==0.9.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.22.3
setuptools==80.8.0
shewchuk==6.10.0
shiboken6==6.9.0
shiboken6-generator==6.9.0
six==1.17.0
smbus==1.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
starlette==0.46.2
symba==2.2.1
TBB==0.2
thefuzz==0.22.1
toml==0.10.2
tomli==2.0.1
tqdm==4.67.1
trove-classifiers==2025.5.9.12
typing-inspection==0.4.1
typing_extensions==4.13.2
uc-micro-py==1.0.3
udiskie==2.5.7
ufw==0.36.2
urllib3==2.4.0
validate-pyproject==0.24.1
virtualenv==20.31.2
-e git+https://github.com/volatilityfoundation/volatility3.git@f5feee842df4a398e84fbade35b75b3e641a38a5#egg=volatility3
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
xcffib==1.5.0
yara-python==4.5.4
yt-dlp==2025.5.22
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
